the one that really got away
so im not sure how maybe its cause im tired and dslyxic but this one actually revert back to a previous version even after ive edited it multiple times today. this does fix the abaility to view a single entry. so now the issue is viewing details after submitting an entry